requirements.txt
# =====================
# Runtime dependencies
# =====================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
textblob==0.17.1
nltk==3.8.1
streamlit==1.29.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4

# =====================
# ML / NLP stack
# =====================
torch==2.1.0
transformers==4.40.0
datasets==2.15.0
scikit-learn==1.3.0
tqdm==4.67.0

# =====================
# Data & visualization
# =====================
pandas==2.1.1
numpy==1.24.3
matplotlib==3.8.0
seaborn==0.13.2
joblib==1.3.2
